requirements.txt
Django>=5.2.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.5.0
